I have added the resource to the repos.
This resource provides an in-depth look at the features of operating systems, and I am sure it will be a great help for anyone looking for information on the subject. and It contains a comprehensive list of questions related to system design, which can be used to prepare for job interviews.